+++ Aoc2020/Day04.cs
@@ -11,13 +11,13 @@ namespace Aoc2020
         public bool IsPart1Complete => true;
         public bool IsPart2Complete => true;
 
-        public int Solve1(IList<string> input)
+        public long Solve1(IList<string> input)
         {
             var passports = ParseInput(input);
             return passports.Count(x => x.ContainsRequiredFields());
         }
 
-        public int Solve2(IList<string> input)
+        public long Solve2(IList<string> input)
         {
             var passports = ParseInput(input);
             return passports.Count(x => x.IsValid());
