+++ Aoc2020/Day22.cs
@@ -9,12 +9,12 @@ namespace Aoc2020
         public bool IsPart1Complete => false;
         public bool IsPart2Complete => false;
 
-        public int Solve1(IList<string> input)
+        public long Solve1(IList<string> input)
         {
             throw new NotImplementedException();
         }
 
-        public int Solve2(IList<string> input)
+        public long Solve2(IList<string> input)
         {
             throw new NotImplementedException();
         }
